chat: move chat commands to their own files
