+++ src/utils/ChatCommands.js
@@ -1,3 +1,5 @@
+import find from 'array-find';
+
 const debug = require('debug')('uwave:chat-commands');
 
 const commands = {};
@@ -25,259 +27,8 @@ export function execute(state, name, args = []) {
   }
 }
 
-// TODO move default commands to some other file!
-import find from 'array-find';
-import { currentUserSelector, userListSelector } from '../selectors/userSelectors';
-
-function findUser(users, username) {
+// Helper to consistently find online users in command handlers.
+export function findUser(users, username) {
   const lname = username.toLowerCase();
   return find(users, o => o.username.toLowerCase() === lname);
 }
-
-import { log } from '../actions/ChatActionCreators';
-import {
-  skipCurrentDJ,
-  moveWaitlistUser, removeWaitlistUser,
-  deleteChatMessagesByUser, deleteAllChatMessages,
-  setUserRole
-} from '../actions/ModerationActionCreators';
-import {
-  joinWaitlist,
-  modClearWaitlist, modLockWaitlist, modUnlockWaitlist
-} from '../actions/WaitlistActionCreators';
-import {
-  djAndWaitlistUsersSelector,
-  waitlistUsersSelector
-} from '../selectors/waitlistSelectors';
-
-const ROLE_MODERATOR = 2;
-const ROLE_MANAGER = 3;
-const hasRole = role => state => currentUserSelector(state).role >= role;
-const isModerator = hasRole(ROLE_MODERATOR);
-const isManager = hasRole(ROLE_MANAGER);
-
-register('help', 'List available commands.', {
-  action: () => (dispatch, getState) => {
-    const available = getCommands();
-    dispatch(log('Available commands:'));
-    Object.keys(available).sort().forEach(name => {
-      const command = available[name];
-      if (canExecute(getState(), command)) {
-        dispatch(log(`/${name} - ${command.description}`));
-      }
-    });
-  }
-});
-
-register('skip', 'Skip the current DJ.', {
-  guard: isModerator,
-  action: (...args) => skipCurrentDJ(args ? args.join(' ') : '[No reason given]')
-});
-
-register('clearchat',
-  'Delete all chat messages. ' +
-  'Pass a username ("/clearchat kool_panda") to only delete messages by that user.',
-  {
-    guard: isModerator,
-    action: (...args) => (dispatch, getState) => {
-      const username = args.join(' ').trim();
-      if (username) {
-        const user = findUser(userListSelector(getState()), username);
-        if (user) {
-          return dispatch(deleteChatMessagesByUser(user._id));
-        }
-        return dispatch(log(`User ${username} is not online right now.`));
-      }
-      return dispatch(deleteAllChatMessages());
-    }
-  }
-);
-
-register(
-  'wladd',
-  'Add a user to the waitlist. Syntax: "/wladd username"',
-  {
-    guard: isModerator,
-    action: username => (dispatch, getState) => {
-      if (!username) {
-        return dispatch(log('Provide a user to add to the waitlist. Syntax: "/wladd username"'));
-      }
-
-      const user = findUser(userListSelector(getState()), username);
-      if (user) {
-        return dispatch(joinWaitlist(user));
-      }
-      return dispatch(log(`User ${username} is not online right now.`));
-    }
-  }
-);
-
-register(
-  'wlremove',
-  'Remove a user from the waitlist. Syntax: "/wlremove username"',
-  {
-    guard: isModerator,
-    action: username => (dispatch, getState) => {
-      if (!username) {
-        return dispatch(log('Provide a user to remove from the waitlist. Syntax: "/wlremove username"'));
-      }
-
-      const user = findUser(
-        djAndWaitlistUsersSelector(getState()),
-        username
-      );
-      if (user) {
-        return dispatch(removeWaitlistUser(user));
-      }
-      return dispatch(log(`User ${username} is not in the waitlist.`));
-    }
-  }
-);
-
-register(
-  'wlclear',
-  'Remove everyone from the waitlist.',
-  { guard: isModerator, action: modClearWaitlist }
-);
-
-register(
-  'wllock',
-  'Lock the waitlist.',
-  { guard: isModerator, action: modLockWaitlist }
-);
-
-register(
-  'wlunlock',
-  'Unlock the waitlist.',
-  { guard: isModerator, action: modUnlockWaitlist }
-);
-
-register(
-  'wlmove',
-  'Move a user to a different position in the waitlist. ' +
-  'Syntax: "/wlmove username position"',
-  {
-    guard: isModerator,
-    action: (username, posStr) => (dispatch, getState) => {
-      if (!username) {
-        return dispatch(log('Provide a user to move in the waitlist. Syntax: "/wlmove username position"'));
-      }
-      const position = parseInt(posStr, 10) - 1;
-      if (!isFinite(position) || position < 0) {
-        return dispatch(log(`Provide a position to move @${username} to. Syntax: "/wlmove username position"`));
-      }
-
-      const user = findUser(
-        waitlistUsersSelector(getState()),
-        username
-      );
-      if (user) {
-        return dispatch(moveWaitlistUser(user, position));
-      }
-      return dispatch(log(`User ${username} is not in the waitlist.`));
-    }
-  }
-);
-
-const roleNames = {
-  user: 0,
-  default: 0,
-  normal: 0,
-  none: 0,
-  special: 1,
-  moderator: 2,
-  mod: 2,
-  manager: 3,
-  admin: 4
-};
-register(
-  'userrole',
-  'Assign a different role to a user. Syntax: "/userrole username role"',
-  {
-    guard: isManager,
-    action: (username, role) => (dispatch, getState) => {
-      if (!username) {
-        return dispatch(log('Provide a user to promote or demote.'));
-      }
-      if (!(role in roleNames)) {
-        return dispatch(log(
-          `Provide a role to promote ${username} to. [user, special, moderator, manager, admin]`
-        ));
-      }
-      const user = findUser(
-        userListSelector(getState()),
-        username
-      );
-      return dispatch(setUserRole(user, roleNames[role]));
-    }
-  }
-);
-
-import ms from 'ms';
-import { mutedUsersSelector } from '../selectors/chatSelectors';
-import { muteUser, unmuteUser } from '../actions/ModerationActionCreators';
-register(
-  'mute',
-  'Mute a user in chat, preventing them from chatting. Syntax: "/mute username duration"',
-  {
-    guard: isModerator,
-    action: (username, duration = '30m') => (dispatch, getState) => {
-      if (!username) {
-        return dispatch(log('Provide a user to mute.'));
-      }
-      const user = findUser(
-        userListSelector(getState()),
-        username
-      );
-      if (!user) {
-        return dispatch(log(`User "${username}" is not online.`));
-      }
-      return dispatch(muteUser(user, ms(`${duration}`)));
-    }
-  }
-);
-
-register(
-  'unmute',
-  'Unmute a user in chat. Syntax: "/unmute username"',
-  {
-    guard: isModerator,
-    action: username => (dispatch, getState) => {
-      if (!username) {
-        return dispatch(log('Provide a user to unmute.'));
-      }
-      const user = findUser(mutedUsersSelector(getState()), username);
-      if (!user) {
-        return dispatch(log(`User "${username}" is not muted.`));
-      }
-      return dispatch(unmuteUser(user));
-    }
-  }
-);
-
-import { banUser } from '../actions/ModerationActionCreators';
-
-register(
-  'ban',
-  'Ban a user. Syntax: "/ban username duration" or "/ban username perma"',
-  {
-    guard: isModerator,
-    action: (username, duration = '24h') => (dispatch, getState) => {
-      if (!username) {
-        return dispatch(log('Provide a user to ban.'));
-      }
-      const user = findUser(
-        userListSelector(getState()),
-        username
-      );
-      if (!user) {
-        return dispatch(log(`User "${username}" is not online.`));
-      }
-      const permanent = duration === 'perma';
-      return dispatch(banUser(user, {
-        duration: ms(`${duration}`),
-        permanent
-      }));
-    }
-  }
-);
