+++ src/utils/commands/help.js
@@ -1,0 +1,15 @@
+import { getCommands, register, canExecute } from '../ChatCommands';
+import { log } from '../../actions/ChatActionCreators';
+
+register('help', 'List available commands.', {
+  action: () => (dispatch, getState) => {
+    const available = getCommands();
+    dispatch(log('Available commands:'));
+    Object.keys(available).sort().forEach(name => {
+      const command = available[name];
+      if (canExecute(getState(), command)) {
+        dispatch(log(`/${name} - ${command.description}`));
+      }
+    });
+  }
+});
