+++ src/utils/commands/index.js
@@ -1,0 +1,6 @@
+import './bans';
+import './chat';
+import './help';
+import './mutes';
+import './staff';
+import './waitlist';
